refactor: restructure server.rs module
